refactor(shape_inference): extract shared test harness

Copilot-Session: 158c8501-8b25-46b5-8b93-ed1844bb9b3c
